+++ react/components/SearchQuery.js
@@ -6,6 +6,7 @@ import searchMetadataQuery from 'vtex.store-resources/QuerySearchMetadataV2'
 import facetsQuery from 'vtex.store-resources/QueryFacetsV2'
 import sponsoredProductsQuery from 'vtex.store-resources/QuerySponsoredProducts'
 import { equals } from 'ramda'
+import { canUseDOM } from 'exenv'
 
 import {
   buildSelectedFacetsAndFullText,
@@ -15,6 +16,30 @@ import {
 import { FACETS_RENDER_THRESHOLD } from '../constants/filterConstants'
 import useRedirect from '../hooks/useRedirect'
 import useSession from '../hooks/useSession'
+
+function getCookie(cname) {
+  if (!canUseDOM) {
+    return null
+  }
+
+  const name = `${cname}=`
+  const decodedCookie = decodeURIComponent(document.cookie)
+  const ca = decodedCookie.split(';')
+
+  for (let i = 0; i < ca.length; i++) {
+    let c = ca[i]
+
+    while (c.charAt(0) === ' ') {
+      c = c.substring(1)
+    }
+
+    if (c.indexOf(name) === 0) {
+      return c.substring(name.length, c.length)
+    }
+  }
+
+  return ''
+}
 
 const DEFAULT_PAGE = 1
 
@@ -155,7 +180,12 @@ const useQueries = (
     getSettings('vtex.store')?.enableFiltersFetchOptimization
 
   const productSearchResult = useQuery(productSearchQuery, {
-    variables,
+    ssr: false,
+    skip: !canUseDOM,
+    variables: {
+      ...variables,
+      variant: getCookie('sp-variant'),
+    },
   })
 
   const { data: { sponsoredProducts } = [] } = useQuery(