+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- `variant` to the search flow.
+
 ## [3.125.0] - 2023-08-24
 
 ### Added